create hook for speed chart settings querystring
